fix(parity): resolve alt-weapon replay divergence and float-fidelity drift (#291)
* feat(diff-verifiers): show rich state diff on divergence

use deepdiff and rich to print a color-coded field-by-field diff
of the full python vs zig state at the first diverging tick.

* chore: add deepdiff dep

* refactor(trace): remove debug-only hashes and aggregate checksums

remove creature_state_hash, projectile_state_hash, *_active_index_sum,
and *_active_index_xor from the trace schema. these were redundant now
that per-entity entries arrays provide full state for comparison.

* fix(zig): match native aim heading fpatan rounding

* fix(python): compute aim heading with native fpatan formula

* fix(zig): align movement and heading precision with native

* docs(float-parity): document x87 PC_53 startup control-word evidence

* feat(trace): show tick number in state diff header

* fix(zig): keep movement and cooldown math in wide precision

* fix(python): narrow fire cooldown updates to f32

* fix(python): enforce f32 projectile spawn and damage precision

* fix(python): narrow creature lifecycle and dead-slide math to f32

* fix(python): match rush spawn health add-order rounding

* fix(python): f32-chain shot jitter heading math

* fix(replay): persist reload_down in packed input flags

* feat(replay): append git sha for non-release replay headers

* fix(ci): drop removed trace hashes and update flag unpacking

* test(parity): align f32 expectation math and refresh projectile snapshot

+++ tests/test_radioactive_perk.py
@@ -4,6 +4,7 @@ from crimson.creatures.runtime import CREATURE_LIFECYCLE_ALIVE, CreaturePool, Cr
 from crimson.creatures.spawn import CreatureFlags
 from crimson.effects import FxQueue
 from crimson.gameplay import GameplayState
+from crimson.math_parity import f32
 from crimson.perks import PerkId
 from crimson.sim.state_types import PlayerState
 from grim.geom import Vec2
@@ -63,7 +64,7 @@ def test_radioactive_kill_awards_base_xp_and_bypasses_death_multipliers() -> Non
     assert player.experience == 112
     assert not result.deaths
     assert creature.hp < 0.0
-    assert_float_close(creature.lifecycle_stage, CREATURE_LIFECYCLE_ALIVE - dt)
+    assert_float_close(creature.lifecycle_stage, CREATURE_LIFECYCLE_ALIVE - float(f32(float(dt))))
     assert fx_queue.count == 1
