+++ tests/test_toxic_avenger_perk.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from crimson.creatures.runtime import CREATURE_LIFECYCLE_ALIVE, CreaturePool, CreatureUpdateOptions
 from crimson.creatures.spawn import CreatureFlags
 from crimson.gameplay import GameplayState
+from crimson.math_parity import f32
 from crimson.perks import PerkId
 from crimson.sim.state_types import PlayerState
 from grim.geom import Vec2
@@ -45,7 +46,7 @@ def test_toxic_avenger_strong_tick_overrides_weak_tick() -> None:
 
     pool.update(dt, options=CreatureUpdateOptions(state=state, players=[player]))
 
-    assert_float_close(creature.hp, 100.0 - dt * 180.0)
+    assert_float_close(creature.hp, 100.0 - float(f32(float(dt))) * 180.0)
 
 
 def test_toxic_avenger_skips_when_player_shielded() -> None:
@@ -66,4 +67,3 @@ def test_toxic_avenger_skips_when_player_shielded() -> None:
     pool.update(0.2, options=CreatureUpdateOptions(state=state, players=[player]))
 
     assert not (creature.flags & CreatureFlags.SELF_DAMAGE_TICK_STRONG)
-
